+++ .github/workflows/ci.yaml
@@ -73,4 +73,3 @@ jobs:
             'core/**/*.rs'
             'literal/**/*.rs'
             'parser/**/*.rs'
-            'ast/asdl_rs.py'
