+++ parser/src/gen/parse.rs
@@ -1,4 +1,4 @@
-// File automatically generated by ast/asdl_rs.py.
+// This file was originally generated from asdl by a python script, but we now edit it manually
 
 impl Parse for ast::StmtFunctionDef {
     fn lex_starts_at(
