+++ scripts/cspell.sh
@@ -1,3 +1,2 @@
 #!/bin/bash
 cspell "ast/**/*.rs" "literal/**/*.rs" "core/**/*.rs" "parser/**/*.rs"
-cspell ast/asdl_rs.py
